requirements.txt
fastapi
uvicorn
pydantic
google-genai
asyncpg
python-dotenv